Rename files to be concistent with its packages
