+++ backend-refactor/app/db/__init__.py
@@ -1,7 +1,7 @@
 """Database package initialization."""
-from app.db.signal_db import load_signals
-from app.db.asset_db import get_assets
-from app.db.measurement_db import get_measurements
+from app.db.signal import load_signals
+from app.db.asset import get_assets
+from app.db.measurement import get_measurements
 
 __all__ = [
     "load_signals", "get_assets", "get_measurements"
