+++ backend-refactor/app/db/measurement.py
