+++ backend-refactor/app/db/signal.py
