+++ backend-refactor/app/schemas/asset.py
