+++ backend-refactor/app/services/__init__.py
@@ -1,8 +1,7 @@
 """Services package."""
-from app.services.asset_service import AssetService
-from app.services.measurement_svc import MeasurementService, get_measurements_for_signals
+from app.services.asset import AssetService
+from app.services.measurement import MeasurementService, get_measurements_for_signals
 
 __all__ = [
-    "AssetService",
-    "MeasurementService", "get_measurements_for_signals"
+    "AssetService", "MeasurementService", "get_measurements_for_signals"
 ]
