+++ backend-refactor/app/services/asset.py
@@ -1,7 +1,7 @@
 """Asset service layer."""
 from typing import List, Dict
-from app.db.asset_db import get_assets
-from app.utils.asset_helper import format_asset_response
+from app.db.asset import get_assets
+from app.utils.asset_utils import format_asset_response
 
 class AssetService:
     """Service for managing assets."""
