+++ backend-refactor/app/utils/__init__.py
@@ -1,5 +1,5 @@
 """Utilities package."""
-from app.utils.asset_helper import format_asset_response, process_asset_data
+from app.utils.asset_utils import format_asset_response, process_asset_data
 from app.utils.date_utils import parse_date, validate_date_range, check_date_range, is_valid_date_range, get_date
 from app.utils.measurement_utils import filter_measurements_by_date, format_measurement
