+++ backend-refactor/app/utils/asset_utils.py
